+++ src/modules/accessories/accessories.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { AccessoriesService } from './accessories.service';
 import { AccessoriesGateway } from './accessories.gateway';
+import { AccessoriesController } from './accessories.controller';
 import { ConfigModule } from '../../core/config/config.module';
 import { LoggerModule } from '../../core/logger/logger.module';
 
@@ -17,6 +18,9 @@ import { LoggerModule } from '../../core/logger/logger.module';
   ],
   exports: [
     AccessoriesService,
+  ],
+  controllers: [
+    AccessoriesController,
   ]
 })
 export class AccessoriesModule { }