+++ test/e2e/plugins.e2e-spec.ts
@@ -1,10 +1,9 @@
 import * as path from 'path';
 import * as fs from 'fs-extra';
-import axios, { AxiosResponse } from 'axios';
+import axios from 'axios';
 import { ValidationPipe, HttpService } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { FastifyAdapter, NestFastifyApplication, } from '@nestjs/platform-fastify';
-import { of } from 'rxjs';
 import { AuthModule } from '../../src/core/auth/auth.module';
 import { PluginsModule } from '../../src/modules/plugins/plugins.module';
 import { HomebridgePlugin } from '../../src/modules/plugins/types';
@@ -138,35 +137,6 @@ describe('PluginController (e2e)', () => {
       }
     });
 
-    expect(res.statusCode).toEqual(200);
-  });
-
-  it('GET /plugins/release/:plugin-name', async () => {
-    const data = {
-      name: 'v1.0.0',
-      body: 'Hello!'
-    };
-
-    const response: AxiosResponse<any> = {
-      data,
-      headers: {},
-      config: { url: 'https://api.github.com/repos/oznu/homebridge-config-ui-x/releases/latest' },
-      status: 200,
-      statusText: 'OK',
-    };
-
-    jest.spyOn(httpService, 'get')
-      .mockImplementationOnce(() => of(response));
-
-    const res = await app.inject({
-      method: 'GET',
-      path: '/plugins/release/homebridge-config-ui-x',
-      headers: {
-        authorization,
-      }
-    });
-
-    expect(res.json()).toEqual({ name: data.name, changelog: data.body });
     expect(res.statusCode).toEqual(200);
   });